#86: Resolve Vue warning about extraneous non-props attributes
Previously, we used a dynamic component for the filter,
and sent it the same props, whether or not it could use
them.  TextSelect has no need of the rows prop, so Vue
gave us this warning.

Using v-if and v-else has the same effect as the dynamic
component, gives us the opportunity to pass different
props to different components, and is probably more
readable anyway.